Fixed Flow, Reworked DB and Adjusted Accordingly
Fixed Flow so that it is now working and added FK in db and changed the
classes to work with the "new" db. Changed name on the UserBO class to
User. Made some changes in template and flow.xhtml to work with the
passing of view parameters. Also added a few attributes to some classes.